--- linode_api/objects/longview/__init__.py
@@ -1,2 +1,0 @@
-from .client import LongviewClient
-from .subscription import LongviewSubscription
