--- linode_api/objects/longview/subscription.py
@@ -1,13 +1,0 @@
-from __future__ import absolute_import
-
-from linode_api.objects import Base, Property
-
-
-class LongviewSubscription(Base):
-    api_endpoint = 'longview/subscriptions/{id}'
-    properties = {
-        "id": Property(identifier=True),
-        "label": Property(),
-        "clients_included": Property(),
-        "price": Property()
-    }
